refactor: renames routes index files
